requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
cloudgenix==5.6.1b2
idna==3.3
pyaml==21.10.1
PyYAML==6.0
requests==2.28.1
schedule==1.1.0
urllib3==1.26.11
websockets==10.3
